backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
python-dotenv>=1.0.0
torch
transformers
accelerate
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-qdrant>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.11.0
sentence-transformers>=3.0.0
python-multipart>=0.0.9
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
